blog/backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.1.1
cryptography==41.0.7
supabase==2.3.4
requests==2.31.0